# api/uspace/uspacedocs_swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for submitting/monitoring jobs to/from an execution machine",
        "title": "Uspace API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8079",
    "basePath": "/api/v1",
    "paths": {
...
        "/resource/download": {
            "get": {
                "description": "Downloads a single resource specified by the access context (must be a file, not a directory).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Download a resource",
                "responses": {
                    "200": {
                        "description": "File downloaded",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad request (e.g., target is a directory or vid is invalid)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to access or stream the resource",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/resource/upload": {
            "post": {
                "description": "Uploads one or more files to a specified resource path from multipart form data.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Upload files",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Files to upload (multiple supported)",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Files uploaded successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request or form error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity (insertion error)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}